--- web/makeathon-2/components/Cloud/index.tsx
@@ -1,3 +1,0 @@
-import Cloud from "./Cloud";
-
-export default Cloud;
