--- web/makeathon-2/styles/Cloud.module.css
@@ -1,4 +1,0 @@
-.cloud {
-    position: absolute;
-    background: linear-gradient(180deg, white, lightgrey);
-}
